🚚 Rename headers extension
