+++ src/lib/udp/multicast_receiver.cpp
@@ -4,7 +4,7 @@
 #include <boost/asio.hpp>
 #include "boost/bind.hpp"
 
-#include "travesim_adapters/udp/multicast_receiver.hh"
+#include "travesim_adapters/udp/multicast_receiver.hpp"
 
 #define NO_FLAGS 0U
